remove compiler warnings about narrowing conversions and unused parameters and variables
(cherry picked from commit 607f424e2b9cf59e9d59068044876019fd6061e5)

Conflicts:
	test/iterator_facade.cpp

+++ test/zip_iterator_test.cpp
@@ -845,6 +845,6 @@ int main( void )
             << "\nNumber of failed tests: " << static_cast<unsigned int>(num_failed_tests)
             << std::endl;
 
-  return num_failed_tests;
+  return static_cast<int>(num_failed_tests);
 }
